Convert Community Hub to server side rendering w/Browserify+Babel.
