+++ kitsune/bundles.py
@@ -263,7 +263,7 @@ PIPELINE_JS = {
             'js/templates/macros.js',
             'js/templates/search-results-list.js',
             'js/templates/search-results.js',
-            'bower/nunjucks/browsers/nunjucks-slim.js',
+            'bower/nunjucks/browser/nunjucks-slim.js',
             'js/nunjucks.js',
             'js/cached_xhr.js',
             'js/search_utils.js',
@@ -293,11 +293,7 @@ PIPELINE_JS = {
     },
     'community-new': {
         'source_filenames': (
-            'bower/jquery/jquery.min.js',
-            'js/libs/jquery.placeholder.js',
-            'bower/underscore/underscore.js',
-            'bower/moment/moment.js',
-            'bower/libs/react.js',
+            'js/community-new.browserify.js',
         ),
         'output_filename': 'community-new-min.js'
     },
