+++ kitsune/community/templates/community/top_contributors_react.html
@@ -19,6 +19,4 @@
     <script type="application/json" name="contributor-data">
         {{ contributors_json|safe }}
     </script>
-    <script type="text/jsx" src="{{ STATIC_URL }}jsx/TopContributors.jsx"></script>
-    <script type="text/jsx" src="{{ STATIC_URL }}jsx/community-new.jsx"></script>
 {% endblock %}
