+++ kitsune/settings.py
@@ -648,6 +648,7 @@ TOWER_ADD_HEADERS = True
 # Django Pipline
 PIPELINE_COMPILERS = (
   'pipeline.compilers.less.LessCompiler',
+  'kitsune.lib.pipeline_compilers.BrowserifyCompiler',
 )
 
 PIPELINE_DISABLE_WRAPPER = True
@@ -660,6 +661,11 @@ PIPELINE_CSS_COMPRESSOR = 'pipeline.compressors.cssmin.CSSMinCompressor'
 PIPELINE_CSSMIN_BINARY = 'cssmin'
 
 PIPELINE_LESS_BINARY = 'lessc'
+
+PIPELINE_BROWSERIFY_BINARY = 'browserify'
+PIPELINE_BROWSERIFY_ARGUMENTS = '-t babelify -t debowerify'
+if DEBUG:
+    PIPELINE_BROWSERIFY_ARGUMENTS += ' -d'
 
 NUNJUCKS_PRECOMPILE_BIN = 'nunjucks-precompile'
