+++ package.json
@@ -11,10 +11,13 @@
     "gulp-nunjucks": "0.2.2"
   },
   "dependencies": {
+    "babelify": "^5.0.4",
     "bower": "1.3.12",
+    "browserify": "^9.0.3",
+    "cssmin": "0.4.3",
+    "debowerify": "^1.2.0",
     "less": "1.7.0",
     "nunjucks": "1.0.5",
-    "uglify-js": "2.4.13",
-    "cssmin": "0.4.3"
+    "uglify-js": "2.4.13"
   }
 }
